+++ src/NimBLELibraryWrapper.h
@@ -121,6 +121,8 @@ public:
 
   bool hasConnectedDevices() override;
 
+  void setDefaultConnectionTimeout(uint16_t timeoutMs) override;
+
 private:
   static void release();
 
@@ -135,6 +137,7 @@ private:
   uint16_t mMaxAdvertisingIntervalTicks = 3200;
   uint16_t mMinConnectionIntervalTicks = 0x6;
   uint16_t mMaxConnectionIntervalTicks = 0x12;
+  uint16_t mDefaultConnectionTimeoutTicks = 200;
 };
 
 } // namespace sensirion::upt::ble_server