+++ CHANGELOG.md
@@ -6,13 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
+
 - Interface to check if devices are connected
+- Interface to configure default connection timeout
 - Documentation for BleServer and interfaces
 
 ## [1.1.0] - 2025-10-01
 
 ### Changed
+
 - Moved to stable UPT Core 1.0.0
 
 ## [1.0.0] - 2025-09-29